Move styles to a separate file
