+++ __docs__/components/gallery/sections/banner-section.tsx
@@ -4,7 +4,7 @@ import Banner from "@khanacademy/wonder-blocks-banner";
 import {HeadingLarge} from "@khanacademy/wonder-blocks-typography";
 
 import ComponentTile from "../component-tile";
-import {styles} from "../component-gallery";
+import {styles} from "../styles";
 
 export default function BannerSection() {
     return (
