+++ __docs__/components/gallery/sections/birthday-picker-section.tsx
@@ -4,7 +4,7 @@ import BirthdayPicker from "@khanacademy/wonder-blocks-birthday-picker";
 import {HeadingLarge} from "@khanacademy/wonder-blocks-typography";
 
 import ComponentTile from "../component-tile";
-import {styles} from "../component-gallery";
+import {styles} from "../styles";
 
 export default function BirthdayPickerSection() {
     return (
