+++ __docs__/components/gallery/sections/breadcrumbs-section.tsx
@@ -8,7 +8,7 @@ import Link from "@khanacademy/wonder-blocks-link";
 import {HeadingLarge} from "@khanacademy/wonder-blocks-typography";
 
 import ComponentTile from "../component-tile";
-import {styles} from "../component-gallery";
+import {styles} from "../styles";
 
 export default function BannerSection() {
     return (
