+++ __docs__/components/gallery/sections/form-section.tsx
@@ -12,7 +12,7 @@ import {
 import {HeadingLarge} from "@khanacademy/wonder-blocks-typography";
 
 import ComponentTile from "../component-tile";
-import {styles} from "../component-gallery";
+import {styles} from "../styles";
 
 export default function BannerSection() {
     const [selectedValues, setSelectedValues] = React.useState<string[]>([]);
