+++ __docs__/components/gallery/sections/icon-button-section.tsx
@@ -6,7 +6,7 @@ import IconButton from "@khanacademy/wonder-blocks-icon-button";
 import {HeadingLarge} from "@khanacademy/wonder-blocks-typography";
 
 import ComponentTile from "../component-tile";
-import {styles} from "../component-gallery";
+import {styles} from "../styles";
 
 export default function IconButtonSection() {
     return (
