+++ __docs__/components/gallery/sections/link-section.tsx
@@ -4,7 +4,7 @@ import Link from "@khanacademy/wonder-blocks-link";
 import {HeadingLarge} from "@khanacademy/wonder-blocks-typography";
 
 import ComponentTile from "../component-tile";
-import {styles} from "../component-gallery";
+import {styles} from "../styles";
 
 export default function LinkSection() {
     return (
