+++ __docs__/components/gallery/sections/popover-section.tsx
@@ -6,7 +6,7 @@ import {Popover, PopoverContent} from "@khanacademy/wonder-blocks-popover";
 import {HeadingLarge} from "@khanacademy/wonder-blocks-typography";
 
 import ComponentTile from "../component-tile";
-import {styles} from "../component-gallery";
+import {styles} from "../styles";
 
 export default function PopoverSection() {
     const [opened, setOpened] = React.useState(true);
