+++ __docs__/components/gallery/sections/progress-spinner-section.tsx
@@ -5,7 +5,7 @@ import {CircularSpinner} from "@khanacademy/wonder-blocks-progress-spinner";
 import {HeadingLarge} from "@khanacademy/wonder-blocks-typography";
 
 import ComponentTile from "../component-tile";
-import {styles} from "../component-gallery";
+import {styles} from "../styles";
 
 export default function ProgressSpinnerSection() {
     return (
