+++ __docs__/components/gallery/sections/tooltip-section.tsx
@@ -5,7 +5,7 @@ import Tooltip from "@khanacademy/wonder-blocks-tooltip";
 import {HeadingLarge} from "@khanacademy/wonder-blocks-typography";
 
 import ComponentTile from "../component-tile";
-import {styles} from "../component-gallery";
+import {styles} from "../styles";
 
 export default function TooltipSection() {
     return (
